[Refactor] - Categorize tx items prepaidCards
