+++ cardstack/src/components/Transactions/TransactionItem.tsx
@@ -6,10 +6,10 @@ import { ERC20Transaction } from './ERC20Transaction';
 import { MerchantClaimTransaction } from './Merchant/MerchantClaimTransaction';
 import { MerchantCreationTransaction } from './Merchant/MerchantCreationTransaction';
 import { MerchantEarnedRevenueTransaction } from './Merchant/MerchantEarnedRevenueTransaction';
-import { PrepaidCardCreatedTransaction } from './PrepaidCardCreatedTransaction';
-import { PrepaidCardPaymentTransaction } from './PrepaidCardPaymentTransaction';
-import { PrepaidCardSplitTransaction } from './PrepaidCardSplitTransaction';
-import { PrepaidCardTransferTransaction } from './PrepaidCardTransferTransaction';
+import { PrepaidCardCreatedTransaction } from './PrepaidCard/PrepaidCardCreatedTransaction';
+import { PrepaidCardPaymentTransaction } from './PrepaidCard/PrepaidCardPaymentTransaction';
+import { PrepaidCardSplitTransaction } from './PrepaidCard/PrepaidCardSplitTransaction';
+import { PrepaidCardTransferTransaction } from './PrepaidCard/PrepaidCardTransferTransaction';
 import { TransactionBaseCustomizationProps } from './TransactionBase';
 import { MerchantEarnedSpendTransaction } from './Merchant/MerchantEarnedSpendTransaction';
 import { MerchantEarnedSpendAndRevenueTransaction } from './Merchant/MerchantEarnedSpendAndRevenueTransaction';
